requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
PyPDF2>=3.0.0
google-generativeai>=0.3.0
typing-extensions>=4.8.0
mistralai
Pillow>=9.0.0